cloud_run/ext_ga4/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.13.0
google-cloud-storage==2.13.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.3
db-dtypes==1.1.1